# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActiveCriticalAlertView": {
        "properties": {
          "alert_code": {
            "title": "Alert Code",
            "type": "string"
          },
          "escalation_deadline_at": {
            "format": "date-time",
            "title": "Escalation Deadline At",
            "type": "string"
          },
          "escalation_delivery_state": {
            "$ref": "#/components/schemas/CriticalAlertDeliveryState"
          },
          "incident_id": {
            "title": "Incident Id",
            "type": "string"
          },
          "primary_deadline_at": {
            "format": "date-time",
            "title": "Primary Deadline At",
            "type": "string"
          },
          "primary_delivery_state": {
            "$ref": "#/components/schemas/CriticalAlertDeliveryState"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          }
        },
        "required": [
          "incident_id",
          "alert_code",
          "recorded_at",
          "primary_delivery_state",
          "escalation_delivery_state",
          "primary_deadline_at",
          "escalation_deadline_at"
        ],
        "title": "ActiveCriticalAlertView",
        "type": "object"
      },
...
      "ApiErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ApiErrorResponse",
        "type": "object"
      },
...
      "CriticalAlertDeliveryState": {
        "enum": [
          "confirmed",
          "timeout",
          "error",
          "unknown"
        ],
        "title": "CriticalAlertDeliveryState",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationalControlAction": {
        "enum": [
          "pause",
          "drain",
          "flatten",
          "halt",
          "rearm"
        ],
        "title": "OperationalControlAction",
        "type": "string"
      },
      "OperationalControlCommandRequest": {
        "additionalProperties": false,
        "description": "Only browser-authored control input; proof fields are forbidden.",
        "properties": {
          "reason_code": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Reason Code",
            "type": "string"
          }
        },
        "required": [
          "reason_code"
        ],
        "title": "OperationalControlCommandRequest",
        "type": "object"
      },
      "OperationalControlMutationResponse": {
        "properties": {
          "action": {
            "$ref": "#/components/schemas/OperationalControlAction"
          },
          "control": {
            "$ref": "#/components/schemas/OperationalControlTransitionView"
          }
        },
        "required": [
          "action",
          "control"
        ],
        "title": "OperationalControlMutationResponse",
        "type": "object"
      },
...
      "OperationsCoordinatorStatus": {
        "enum": [
          "active",
          "absent",
          "expired",
          "unknown"
        ],
        "title": "OperationsCoordinatorStatus",
        "type": "string"
      },
      "OperationsCoordinatorView": {
        "properties": {
          "fencing_generation": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fencing Generation"
          },
          "lease_expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Expires At"
          },
          "owner_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "status": {
            "$ref": "#/components/schemas/OperationsCoordinatorStatus"
          }
        },
        "required": [
          "status",
          "owner_id",
          "fencing_generation",
          "lease_expires_at"
        ],
        "title": "OperationsCoordinatorView",
        "type": "object"
      },
...
      "OperationsOverviewResponse": {
        "properties": {
          "active_alerts": {
            "items": {
              "$ref": "#/components/schemas/ActiveCriticalAlertView"
            },
            "maxItems": 512,
            "title": "Active Alerts",
            "type": "array"
          },
          "as_of": {
            "format": "date-time",
            "title": "As Of",
            "type": "string"
          },
          "control": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OperationalControlTransitionView"
              },
              {
                "type": "null"
              }
            ]
          },
          "control_history": {
            "items": {
              "$ref": "#/components/schemas/OperationalControlTransitionView"
            },
            "maxItems": 512,
            "title": "Control History",
            "type": "array"
          },
          "coordinator": {
            "$ref": "#/components/schemas/OperationsCoordinatorView"
          },
          "current_risk_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdvancedRiskAssessmentView"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_risk_assignment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AdvancedRiskAssignmentView"
              },
              {
                "type": "null"
              }
            ]
          },
          "environment": {
            "$ref": "#/components/schemas/OperationsEnvironmentView"
          },
          "readiness": {
            "$ref": "#/components/schemas/OperationsReadinessView"
          }
        },
        "required": [
          "as_of",
          "environment",
          "readiness",
          "coordinator",
          "control",
          "control_history",
          "current_risk_assignment",
          "current_risk_assessment",
          "active_alerts"
        ],
        "title": "OperationsOverviewResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "LocalOperatorSession": {
        "description": "URL-safe, server-issued local session value",
        "in": "cookie",
        "name": "aqt_local_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Safety-first quantitative trading control and query API.",
    "title": "AutoQuantTrader API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/operations/accounts/{account_id}": {
      "get": {
        "operationId": "operations_overview_api_v1_operations_accounts__account_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 32,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "X-Csrf-Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsOverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "LocalOperatorSession": []
          }
        ],
        "summary": "Operations Overview",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/operations/accounts/{account_id}/control/{action}": {
      "post": {
        "operationId": "mutate_control_api_v1_operations_accounts__account_id__control__action__post",
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{0,63}$",
              "title": "Account Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "action",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/OperationalControlAction"
            }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 32,
              "pattern": "^[A-Za-z0-9_-]+$",
              "title": "X-Csrf-Token",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$",
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OperationalControlCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationalControlMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "LocalOperatorSession": []
          }
        ],
        "summary": "Mutate Control",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}